user_profile/tokens.py
from django.contrib.auth.tokens import PasswordResetTokenGenerator


class EmailVerificationTokenGenerator(PasswordResetTokenGenerator):
    def _make_hash_value(self, user, timestamp):
        return str(user.pk) + str(timestamp) + str(user.profile.is_email_verified)


email_verification_token = EmailVerificationTokenGenerator()
