dev-requirements.txt
autopep8
django-debug-toolbar
django-rosetta
django-extensions
Werkzeug
coverage
pycodestyle
black